[WIP] Extracted contract test utils to dev-utils package
