+++ packages/contracts/test/newsroom.ts
@@ -1,7 +1,7 @@
 import * as chai from "chai";
 import { configureChai } from "@joincivil/dev-utils";
 
-import { events, NEWSROOM_ROLE_EDITOR, NEWSROOM_ROLE_REPORTER, REVERTED } from "../utils/constants";
+import { events, NEWSROOM_ROLE_EDITOR, NEWSROOM_ROLE_REPORTER, REVERTED } from "./utils/constants";
 import { findEvent, idFromEvent, is0x0Address, timestampFromTx } from "./utils/contractutils";
 
 const Newsroom = artifacts.require("Newsroom");
