+++ packages/contracts/tsconfig.json
@@ -11,6 +11,7 @@
     "./migrations/**/*.ts",
     "./migrations/**/*.js",
     "./scripts/**.ts",
-    "../../node_modules/web3-typescript-typings/index.d.ts"
+    "../../node_modules/web3-typescript-typings/index.d.ts",
+    "../dev-utils/src/test/test.d.ts"
   ]
 }
