+++ packages/multisig/tsconfig.json
@@ -7,6 +7,7 @@
   "include": [
     "./test/**/*.ts",
     "./migrations/**/*.ts",
-    "../../node_modules/web3-typescript-typings/index.d.ts"
+    "../../node_modules/web3-typescript-typings/index.d.ts",
+    "../dev-utils/src/test/test.d.ts"
   ]
 }
