+++ packages/utils/src/index.ts
@@ -1,2 +1,3 @@
 import "./rxjs";
 export * from "./language";
+export * from "./ethereum";
